Implementation of Global Exception Handling.

+++ src/main/java/com/learn/employee_manager/service/EmployeeService.java
@@ -1,11 +1,10 @@
 package com.learn.employee_manager.service;
 
+import com.learn.employee_manager.exception.ResourceNotFoundException;
 import com.learn.employee_manager.model.Employee;
 import com.learn.employee_manager.repository.EmployeeRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 @Service
@@ -19,26 +18,28 @@ public class EmployeeService {
         return employeeRepository.findAll();
     }
 
-    public void addEmployee(Employee employee){
+    public Employee addEmployee(Employee employee){
         employeeRepository.save(employee);
+        return employee;
     }
 
     public Employee updatedEmployee(int id, Employee employee){
 
-        Employee existingEmployee = employeeRepository.findById(id).orElse(null);
+        Employee existingEmployee = employeeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Employee not found with id: " + id));
 
+        existingEmployee.setName(employee.getName());
+        existingEmployee.setEmail(employee.getEmail());
 
-        if (existingEmployee != null){
-            existingEmployee.setName(employee.getName());
-            existingEmployee.setEmail(employee.getEmail());
+        return employeeRepository.save(existingEmployee);
 
-            return employeeRepository.save(existingEmployee);
-        }
-
-        return null;
     }
 
     public void deleteById(int id){
+
+        if (!employeeRepository.existsById(id)){
+            throw new ResourceNotFoundException("Employee not found with id: " + id);
+        }
         employeeRepository.deleteById(id);
     }
 }
